app/requirements.txt
fastapi
uvicorn
typing
redis
aiofiles
loguru
httpx[http2]
google-cloud-firestore
firebase_admin

bcrypt==4.2.0
python-jose==3.3.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
typing-inspect==0.9.0
pydantic==2.8.2
pydantic-settings==2.4.0
loguru==0.7.2
fastapi-restful==0.6.0
bson
# opencv-contrib-python
numpy
opencv-python-headless
imutils
ultralytics
lapx
python-dotenv